* debugger/mcs51/cmd.c, * debugger/mcs51/sdcdb.c, * debugger/mcs51/simi.c, * debugger/mcs51/symtab.c, * src/SDCCargs.h, * src/version.awk: cosmetic changes * src/pic14/gen.c (genDataPointerSet), * src/pic14/glue.c (emitIvals), * src/SDCCsystem.c: fixed warnings * src/SDCCerr.c (setWarningDisabled): check errNum * src/SDCCpeeph.c: added optimizeReturn * src/mcs51/peeph.def: disabled 237.x, fixed bug 1653651 * src/SDCCglobl.h: added options.peepReturn * src/SDCCmain.c (parseCmdLine): added --peep-return & --no-peep-return * support/scripts/resource.rc: replaced include "winres.h" with "winresrc.h" * support/scripts/winres.h: deleted
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@6230 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/pic14/glue.c
@@ -771,9 +771,9 @@ emitIvals(struct dbuf_s *oBuf, symbol *sym, initList *list, long lit, int size)
         // initialize as zero
         for (i = 0; i < size; i++) {
             if (in_code) {
-                dbuf_printf (oBuf, "\tretlw 0x%02x\n", lit & 0xff);
+                dbuf_printf (oBuf, "\tretlw 0x%02x\n", (int)(lit & 0xff));
             } else {
-                dbuf_printf (oBuf, "%s0x%02x", (i == 0) ? "" : ", ", lit & 0xff);
+                dbuf_printf (oBuf, "%s0x%02x", (i == 0) ? "" : ", ", (int)(lit & 0xff));
             }
             lit >>= 8;
         } // for
